fix(samples): finance data
