+++ samples/finance/server/project.json
@@ -4,8 +4,15 @@
   "sourceRoot": "samples/finance/server/src",
   "projectType": "application",
   "targets": {
+    "generate-data": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "npx tsx samples/finance/server/tools/generate-ingredients.ts"
+      }
+    },
     "build": {
       "executor": "@nx/esbuild:esbuild",
+      "dependsOn": ["generate-data"],
       "outputs": ["{options.outputPath}"],
       "defaultConfiguration": "production",
       "options": {
@@ -16,6 +23,13 @@
         "main": "samples/finance/server/src/main.ts",
         "tsConfig": "samples/finance/server/tsconfig.app.json",
         "generatePackageJson": true,
+        "assets": [
+          {
+            "input": "samples/finance/server/src",
+            "glob": "ingredients.json",
+            "output": "samples/finance/server/src"
+          }
+        ],
         "esbuildOptions": {
           "sourcemap": true,
           "outExtension": {
